game: tidy file structure
