+++ src/meson.build
@@ -159,8 +159,9 @@ common_sources = [
   'trx/game/events.c',
   'trx/game/fader.c',
   'trx/game/fmv.c',
-  'trx/game/game.c',
+  'trx/game/game/control.c',
   'trx/game/game/draw.c',
+  'trx/game/game/state.c',
   'trx/game/game_buf.c',
   'trx/game/game_flow/common.c',
   'trx/game/game_flow/inventory.c',
