+++ src/trx/game/game/draw.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void Game_FadeToBlack(int32_t duration);
+void Game_DrawFade(void);
+void Game_Draw(bool draw_overlay);
